Eliminate bad menuconfig combinations
Offer only CBFS_PREFIX "fallback" or "normal", as those names
are hard-coded in the bootblock.

Only with the choice "fallback" user is provided with the option to
select bootblock behaviour.

With choice "normal" build runs in image update mode and user has to
provide a template (with bootblock and fallback) as build/coreboot.rom.

Change-Id: Ie89c2f89245beec70ebb14326437d71699c367b8
Signed-off-by: Kyösti Mälkki <kyosti.malkki@gmail.com>